requirements.txt
boto3==1.34.69
PyJWT==2.8.0
python-dotenv==1.0.1
